combine book search to listings page and updated some styling
