+++ create_account.html
@@ -35,7 +35,7 @@
         </header>
 
         <!-- account information section -->
-        <section class="account_info">
+        <section class="account-info">
             <div class="container">
                 <form action="#">
                     <label id="university" for="university">University</label><br>
