+++ index.html
@@ -37,8 +37,8 @@
         <!-- index page buttons -->
         <section class="index-page-buttons">
             <div class="container">
-                <button type="button" id="list-book"><a href="list_book.html">List Book for Exchange or Donation</a></button>
-                <button type="button" id="personal-listings"><a href="personal_listings.html">View Book Listings</a></button>
+                <button type="button" class="submit-button" id="list-book"><a href="list_book.html">List Book</a></button>
+                <button type="button" class="submit-button" id="book-listings"><a href="book_listings.html">View Book Listings</a></button>
             </div>
         </section>
     </body>
