+++ js/index.js
@@ -1,8 +1,8 @@
 import { checkUserLoggedIn } from "./requests.js";
-import { loginPageFilename, listBookPageFilename, personalListingsPageFilename } from "./vars_and_helpers.js";
+import { loginPageFilename, listBookPageFilename, bookListingsPageFilename } from "./vars_and_helpers.js";
 
 const listBookButton = $("button#list-book");
-const personalListingsButton = $("button#personal-listings");
+const bookListingsButton = $("button#book-listings");
 
 
 $(document).ready(() => {
@@ -14,12 +14,13 @@ $(document).ready(() => {
             sessionStorage.setItem("pageAfterLogin", listBookPageFilename);
         }
     });
-    personalListingsButton.click(e => {
+    
+    bookListingsButton.click(e => {
         // Checks if user is logged in. If not, the user is redirected to the login page filename.
         if (!checkUserLoggedIn()) {
             e.preventDefault();
             window.location = loginPageFilename;
-            sessionStorage.setItem("pageAfterLogin", personalListingsPageFilename);
+            sessionStorage.setItem("pageAfterLogin", bookListingsPageFilename);
         }
     });
 });
