+++ list_book.html
@@ -35,7 +35,7 @@
         </header>
 
         <!-- book information section -->
-        <section class="book_info">
+        <section class="book-info">
             <div class="container">
                 <form action="#">
                     <label id="isbn" for="isbn">ISBN</label><br>
