+++ main.css
@@ -13,6 +13,17 @@ header {
     white-space: pre;
 }
 
+header, section {
+    display: block;
+    font-family: 'Work Sans', sans-serif;
+    font-size: 20px;
+    font-weight: 100;
+    color: #232323;
+    width: fit-content;
+    position: left;
+    margin-left: 5%;
+}
+
 h1 {
     font-family: 'Work Sans', sans-serif;
     font-size: 57px;
@@ -21,7 +32,16 @@ h1 {
     width: fit-content;
     margin-top: 80px;
     margin-bottom: 75px;
-    margin-left: 5%;
+}
+
+h2 {
+    font-family: 'Work Sans', sans-serif;
+    font-size: 30px;
+    font-weight: 100;
+    color: #5A5A5A;
+    width: fit-content;
+    margin-top: 80px;
+    margin-bottom: 15px;
 }
 
 .navbar {
@@ -74,6 +94,7 @@ li a {
   width: 300px;
   height: 50px;
   border-radius: 10px;
+  margin-top: 15px;
   font-family: 'Work Sans', sans-serif;
   font-size: 23px;
   font-weight: 100;
@@ -82,25 +103,18 @@ li a {
   box-shadow:  0px 5px 10px 0px rgba(0, 0, 0, 0.4);
 }
 
-.submit-button a {
+.index-page-buttons .submit-button {
+    display: inline-block;
+    margin-right: 10px;
+}
+
+.submit-button a, button a {
     display: block;
     text-decoration: none;
     color: #ededed;
 }
 
-.account_info, .book_info, .search_info {
-    display: block;
-    font-family: 'Work Sans', sans-serif;
-    font-size: 20px;
-    font-weight: 100;
-    color: #232323;
-    width: fit-content;
-    position: left;
-    margin-top: 5%;
-    margin-left: 5%;
-}
-
-.account_info input, .book_info input, .book_info select, .book_info textarea, .search_info input {
+.account-info input, .book-info input, .book-info select, .book-info textarea, .search-info input {
     width: 500px;
     height: 40px;
     border-radius: 4px;
@@ -114,17 +128,13 @@ li a {
     font-weight: 100;
 }
 
-.book_info textarea {
+.book-info textarea {
     height: 100px;
 }
 
-.book_info input[type="radio"] {
+.book-info input[type="radio"] {
     width: 20px;
     height: 15px;
-}
-
-.index-page-buttons a {
-    color: #125ccb;
 }
 
 .reset-password-link {
@@ -192,7 +202,7 @@ li a {
   margin-bottom: 1%;
 }
 
-.collection-list{
+.collection-list, .search-results-list {
     position: relative;
     margin-top: 2%;
     width: 100%;
@@ -204,4 +214,3 @@ li a {
     margin-bottom: 2%;
     margin-left: 4%;
 }
-
